+++ CHANGES.md
@@ -10,7 +10,7 @@
 * [#52](https://github.com/serokell/universum/issues/52):
   Add `maybeToMonoid :: Monoid m => Maybe m -> m`.
 * Remove `Symbol`-related types for sure.
-* Return back seems to be useful function `guardM` removed in `0.3`.
+* Return back seems to be useful function `guardM` removed in `v0.3`.
 * Add `notElem` for `NonTrivialContainer`.
 
 0.6.1